+++ example/lib/modules/home/controller/home_controller.dart
@@ -1,0 +1,5 @@
+﻿import 'package:get/get.dart';
+
+class HomeController extends GetxController {
+  HomeController();
+}