+++ example/lib/modules/login/bindings/login_bindings.dart
@@ -1,0 +1,9 @@
+﻿import 'package:fuk_ui_kit_sample/modules/login/controller/login_controller.dart';
+import 'package:get/get.dart';
+
+class LoginBinding implements Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<LoginController>(() => LoginController(), fenix: true);
+  }
+}