+++ example/lib/modules/login/controller/login_controller.dart
@@ -1,0 +1,5 @@
+﻿import 'package:get/get.dart';
+
+class LoginController extends GetxController {
+  LoginController();
+}