+++ lib/src/components/layout/sidebar.dart
@@ -5,6 +5,7 @@ class FukSidebar extends StatefulWidget {
   final List<SideBarItems> items;
   final List<SideBarItems>? bottomItems;
   final String? iconImage;
+  final Color? backgroundColor;
   final Icon? icon;
 
   const FukSidebar({
@@ -13,6 +14,7 @@ class FukSidebar extends StatefulWidget {
     this.bottomItems,
     this.iconImage,
     this.icon,
+    this.backgroundColor,
   });
 
   @override
@@ -29,7 +31,7 @@ class FukSidebarState extends State<FukSidebar> {
       height: MediaQuery.of(context).size.height,
       padding: const EdgeInsets.all(8),
       decoration: BoxDecoration(
-        color: Theme.of(context).colorScheme.surface,
+        color: widget.backgroundColor ?? Theme.of(context).colorScheme.surface,
         boxShadow: [
           BoxShadow(
             color: Theme.of(context).shadowColor.withOpacity(0.12),